refactor(ngx-oauth2-access-token): move programming logic to match the level
