+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/errors.ts
@@ -2,5 +2,7 @@ export * from './errors/callback.errors';
 export * from './errors/oauth2-client.errors';
 export * from './errors/scope.errors';
 export * from './errors/state-action.errors';
-export * from './errors/storages.errors';
+
 export * from './errors/access-token.errors';
+export * from './errors/authorization-code.errors';
+export * from './errors/id-token.errors';
