+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/errors/access-token.errors.ts
@@ -1,8 +1,27 @@
-export class UpdateToAccessTokenResponseListenerError extends Error {
-  constructor(name: string, cause: unknown) {
-    super(`Error on '${name}'.`, {
-      cause,
-    });
+export class AccessTokenNotFoundError extends Error {
+  constructor(name: string, options?: ErrorOptions) {
+    super(`Access token of '${name}' is not found.`, options);
+    this.name = this.constructor.name;
+  }
+}
+
+export class AccessTokenExpiredError extends Error {
+  constructor(name: string, options?: ErrorOptions) {
+    super(`Access token of '${name}' has expired.`, options);
+    this.name = this.constructor.name;
+  }
+}
+
+export class RefreshTokenNotFoundError extends Error {
+  constructor(name: string, options?: ErrorOptions) {
+    super(`Refresh token of '${name}' is not found.`, options);
+    this.name = this.constructor.name;
+  }
+}
+
+export class RefreshTokenExpiredError extends Error {
+  constructor(name: string, options?: ErrorOptions) {
+    super(`Refresh token of '${name}' has expired.`, options);
     this.name = this.constructor.name;
   }
 }
