+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/errors/authorization-code.errors.ts
@@ -1,0 +1,13 @@
+export class StateNotFoundError extends Error {
+  constructor(message = 'State not found.', options?: ErrorOptions) {
+    super(message, options);
+    this.name = this.constructor.name;
+  }
+}
+
+export class StateExpiredError extends Error {
+  constructor(message = 'State has expired.', options?: ErrorOptions) {
+    super(message, options);
+    this.name = this.constructor.name;
+  }
+}
