+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/types/storages.types.ts
@@ -2,10 +2,7 @@ import { Observable } from 'rxjs';
 
 import { AccessTokenResponse, JwtTokenType } from './standard.types';
 
-export type StoredAccessTokenResponse = Omit<
-  AccessTokenResponse,
-  'expires_in' | 'refresh_token'
-> & {
+export type StoredAccessTokenResponse = AccessTokenResponse & {
   expires_at: number;
 };
 
@@ -26,6 +23,13 @@ export interface KeyValuePairStorage {
   loadItem<T = unknown>(key: string): Promise<T | null>;
   storeItem<T = unknown>(key: string, value: T): Promise<T>;
   removeItem<T = unknown>(key: string): Promise<T | null>;
+
+  /**
+   * Return a _multicast observable_ for changing value of the given `key`. The
+   * `null` value means the key was removed.
+   */
   watchItem<T = unknown>(key: string): Observable<T | null>;
+
+  /** Return all keys from the storage. */
   keys(): Promise<string[]>;
 }
